refactor: ensure task manager is retry-able and test-able
By passing the tasks to the manager at construction time, and not
immediately starting it, we can improve how it will be handled in the
exporter's open phase if that gets retried, and make sure it is retried
in a safe way (e.g. do not resubmit the same task). That way, we ensure
there is not the same task running twice which could cause concurrency
issues.

Additionally, this will make it easier to unit and integration test,
because we can now control when and how the tasks are executed, and have
better introspection into the background task manager.